dividing files
