+++ dnsext-svcb/DNS/SVCB/Value.hs
@@ -1,0 +1,5 @@
+module DNS.SVCB.Value where
+
+import DNS.Types
+
+type SvcParamValue = Opaque
