+++ dnsext-svcb/dnsext-svcb.cabal
@@ -25,6 +25,9 @@ library
 
     other-modules:
         DNS.SVCB.Imports
+        DNS.SVCB.Key
+        DNS.SVCB.Value
+        DNS.SVCB.Params
 
     default-language: Haskell2010
     ghc-options:      -Wall
